+++ apps/ledger-live-desktop/src/newArch/Collectibles/Nfts/Collection/index.tsx
@@ -1,151 +1,82 @@
-import React, { useMemo } from "react";
-import { useNftCollectionModel } from "./useNftCollectionModel";
-import { Box, Icons, Flex } from "@ledgerhq/react-ui";
-import TableContainer from "~/renderer/components/TableContainer";
-import TableHeader from "../../components/Collection/TableHeader";
-import { Account, ProtoNFT } from "@ledgerhq/types-live";
-import { useNftMetadata } from "@ledgerhq/live-nft-react";
-import TableRow from "../../components/Collection/TableRow";
-import EmptyCollection from "../../components/Collection/EmptyCollection";
-import { CollectibleTypeEnum } from "../../types/Collectibles";
-import Button from "~/renderer/components/Button";
+import React, { memo } from "react";
+import useNftCollectionModel from "./useNftCollectionModel";
 import { useTranslation } from "react-i18next";
-import ShowMore from "../../components/Collection/ShowMore";
-import { FieldStatus } from "../../types/DetailDrawer";
-import CollectionContextMenu from "LLD/components/ContextMenu/CollectibleContextMenu";
-import HeaderActions from "../../components/Collection/HeaderActions";
-import { TableHeaderProps, TableHeaderTitleKey as TitleKey } from "../../types/Collection";
-
-export type NftsInTheCollections = {
-  contract: string;
-  nft: ProtoNFT;
-  nftsNumber: number;
-  onClick: (collectionAddress: string) => void;
-};
+import TrackPage from "~/renderer/analytics/TrackPage";
+import Box from "~/renderer/components/Box/Box";
+import Button from "~/renderer/components/Button";
+import Text from "~/renderer/components/Text";
+import { Flex, Icons } from "@ledgerhq/react-ui";
+import { Skeleton, Media, CollectionName, TableLayout } from "../../components";
+import TokensList from "../Gallery/components/TokensList";
+import OperationsList from "~/renderer/components/OperationsList";
 
 type ViewProps = ReturnType<typeof useNftCollectionModel>;
 
-type NftItemProps = {
-  contract: string;
-  tokenId: string;
-  currencyId: string;
-  numberOfNfts: number;
-  account: Account;
-  onClick: (collectionAddress: string) => void;
-};
-
-type Props = {
-  account: Account;
-};
-
-const NftItem: React.FC<NftItemProps> = ({
-  contract,
-  tokenId,
-  currencyId,
-  account,
-  numberOfNfts,
-  onClick,
-}) => {
-  const { metadata, status } = useNftMetadata(contract, tokenId, currencyId);
-  const isLoading = status === FieldStatus.Loading;
-
-  return (
-    <CollectionContextMenu
-      collectionName={metadata?.tokenName}
-      collectionAddress={contract}
-      account={account}
-      typeOfCollectible={CollectibleTypeEnum.NFT}
-    >
-      <TableRow
-        isLoading={isLoading}
-        tokenName={metadata?.tokenName || metadata?.nftName || "-"}
-        numberOfNfts={numberOfNfts}
-        onClick={() => onClick(contract)}
-        media={{
-          isLoading: isLoading,
-          useFallback: true,
-          contentType: "image",
-          uri: metadata?.medias.preview.uri,
-          mediaType: metadata?.medias.preview.mediaType,
-        }}
-      />
-    </CollectionContextMenu>
-  );
-};
-
 function View({
-  nftsInTheCollection,
+  isLoading,
   account,
-  displayShowMore,
-  onOpenGallery,
-  onReceive,
-  onOpenCollection,
-  onShowMore,
+  collectionAddress,
+  nfts,
+  metadata,
+  filterOperation,
+  onSend,
 }: ViewProps) {
   const { t } = useTranslation();
-  const hasNfts = nftsInTheCollection.length > 0;
-
-  const actions = useMemo(() => {
-    return hasNfts
-      ? [
-          {
-            element: (
-              <HeaderActions textKey="NFT.collections.receiveCTA">
-                <Icons.ArrowDown size="S" />
-              </HeaderActions>
-            ),
-            action: onReceive,
-          },
-          {
-            element: <HeaderActions textKey="NFT.collections.galleryCTA" />,
-            action: onOpenGallery,
-          },
-        ]
-      : [];
-  }, [hasNfts, onReceive, onOpenGallery]);
-
-  const tableHeaderProps: TableHeaderProps = {
-    titleKey: TitleKey.NFTCollections,
-    actions,
-  };
 
   return (
-    <Box>
-      <TableContainer id="tokens-list" mb={50}>
-        <TableHeader {...tableHeaderProps} />
-        {hasNfts ? (
-          nftsInTheCollection.map(item => (
-            <NftItem
-              key={`${item.contract}-${item.nft.tokenId}`}
-              contract={item.contract}
-              tokenId={item.nft.tokenId}
-              account={account}
-              currencyId={item.nft.currencyId}
-              numberOfNfts={item.nftsNumber}
-              onClick={onOpenCollection}
-            />
-          ))
-        ) : (
-          <EmptyCollection
-            currencyName={account.currency.name}
-            collectionType={CollectibleTypeEnum.NFT}
-          >
-            <Button small primary onClick={onReceive} icon>
-              <Flex alignItems={"center"}>
-                <Icons.ArrowDown size="XS" />
-                <Box>{t("NFT.collections.receiveCTA")}</Box>
-              </Flex>
-            </Button>
-          </EmptyCollection>
-        )}
-        {displayShowMore && <ShowMore onShowMore={onShowMore} />}
-      </TableContainer>
-    </Box>
+    <>
+      <TrackPage category="Page" name="NFT Specific Collection" />
+      <Flex flexDirection={"column"} mb={16} width={"100%"}>
+        <Flex flex={1} justifyContent={"space-between"}>
+          <Box horizontal alignItems="center" mb={16}>
+            <Skeleton width={40} minHeight={40} show={isLoading}>
+              <Media
+                size={40}
+                uri={metadata?.medias.preview.uri}
+                tokenId={metadata?.tokenId}
+                mediaType={metadata?.medias.preview.mediaType}
+                isLoading={isLoading}
+                contentType="image"
+                useFallback
+              />
+            </Skeleton>
+            <Box ml={3}>
+              <Skeleton width={93} barHeight={6} minHeight={24} show={isLoading}>
+                <Text ff="Inter|Regular" color="palette.text.shade60" fontSize={2}>
+                  {t("NFT.gallery.collection.header.contract", {
+                    contract: collectionAddress,
+                  })}
+                </Text>
+                <Skeleton width={93} barHeight={6} minHeight={24} show={isLoading}>
+                  <Text ff="Inter|SemiBold" color="palette.text.shade100" fontSize={22}>
+                    <CollectionName nft={nfts[0]} />
+                  </Text>
+                </Skeleton>
+              </Skeleton>
+            </Box>
+          </Box>
+          <Button small primary icon onClick={onSend}>
+            <Box horizontal flow={1} alignItems="center">
+              <Icons.ArrowUp size={"S"} />
+              <Box>{t("NFT.gallery.collection.header.sendCTA")}</Box>
+            </Box>
+          </Button>
+        </Flex>
+        <TableLayout />
+        {account && <TokensList account={account} nfts={nfts} />}
+        <OperationsList
+          account={account}
+          title={t("NFT.gallery.collection.operationList.header")}
+          filterOperation={collectionAddress ? filterOperation : undefined}
+          t={t}
+        />
+      </Flex>
+    </>
   );
 }
 
-const NftCollection: React.FC<Props> = ({ account }) => {
-  return <View {...useNftCollectionModel({ account })} account={account} />;
+const NftCollection: React.FC = () => {
+  return <View {...useNftCollectionModel()} />;
 };
 
-export default NftCollection;
+export default memo(NftCollection);
