+++ apps/ledger-live-desktop/src/newArch/Collectibles/Nfts/DetailDrawer/Actions.tsx
