+++ apps/ledger-live-desktop/src/newArch/Collectibles/Nfts/DetailDrawer/index.tsx
