+++ apps/ledger-live-desktop/src/newArch/Collectibles/__integration__/shared.tsx
@@ -1,9 +1,10 @@
 import { genAccount } from "@ledgerhq/coin-framework/lib/mocks/account";
 import React from "react";
 import { Switch, Route, withRouter } from "react-router";
-import NftCollection from "../Nfts/Collection";
 import NFTGallery from "~/renderer/screens/nft/Gallery";
 import NftBreadCrumb from "LLD/components/BreadCrumb";
+import NftCollection from "../Nfts/Collection";
+import NftCollections from "../Nfts/Collections";
 
 export const account = genAccount("ethereum1", {
   withNft: true,
@@ -12,8 +13,9 @@ export const account = genAccount("ethereum1", {
 
 const NftCollectionNavigation = () => (
   <Switch>
-    <Route exact path="/" render={() => <NftCollection account={account} />} />
+    <Route exact path="/" render={() => <NftCollections account={account} />} />
     <Route path="/account/:id/nft-collection" render={() => <NFTGallery />} />
+    <Route path="/account/:id/nft-collection/:collectionAddress" render={() => <NftCollection />} />
   </Switch>
 );
 
@@ -28,6 +30,6 @@ export const NftCollectionTest = withRouter(NftCollectionTestBase);
 export const NoNftCollectionTest = withRouter(() => (
   <>
     <NftBreadCrumb />
-    <NftCollection account={genAccount("ethereum1")} />
+    <NftCollections account={genAccount("ethereum1")} />
   </>
 ));
