+++ apps/ledger-live-desktop/src/renderer/screens/nft/Gallery/Collection.tsx
@@ -9,7 +9,7 @@ import { openModal } from "~/renderer/actions/modals";
 import styled from "styled-components";
 import Media from "~/renderer/components/Nft/Media";
 import IconSend from "~/renderer/icons/Send";
-import TokensList from "./TokensList";
+import TokensList from "LLD/Collectibles/Nfts/Gallery/components/TokensList";
 import Spinner from "~/renderer/components/Spinner";
 import Box from "~/renderer/components/Box";
 import Button from "~/renderer/components/Button";
@@ -132,9 +132,7 @@ const Collection = () => {
         </Button>
       </Box>
       <GridListToggle />
-      {account && (
-        <TokensList account={account} nfts={slicedNfts} onHideCollection={onCollectionHide} />
-      )}
+      {account && <TokensList account={account} nfts={slicedNfts} />}
 
       <Footer ref={listFooterRef}>
         {nfts.length > maxVisibleNFTs && (
