+++ apps/ledger-live-desktop/src/renderer/screens/nft/Gallery/TokensList/Item.tsx
@@ -13,7 +13,7 @@ import { useNftMetadata } from "@ledgerhq/live-nft-react";
 import NFTContextMenu from "~/renderer/components/ContextMenu/NFTContextMenu";
 import { State } from "~/renderer/reducers";
 import { IconsLegacy } from "@ledgerhq/react-ui";
-import NftDetailDrawer from "~/newArch/Collectibles/Nfts/Gallery/components/DetailDrawer/index";
+import NftDetailDrawer from "~/newArch/Collectibles/Nfts/DetailDrawer/index";
 
 const Wrapper = styled(Card)`
   &.disabled {
